pyproject.toml
[project]
name = "mqtt-client"
version = "0.1.0"
description = "Proof of concept MQTT client using client certificates"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.0.1",
]
authors = [
    { name = "Your Name" }
]
keywords = ["mqtt", "ssl", "tls", "iot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: MQTT",
    "Topic :: Security :: Cryptography",
]

[project.urls]
Repository = "https://github.com/username/mqtt-client"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mqtt_pub.py", "mqtt_sub.py"]
